[FIX] chat: fold conversation on mobile when redirecting to webpage
Before this commit:

When the chatbot redirects a visitor to a webpage on mobile, the chat window
remains open, covering the redirected page. Making it difficult for the
visitor to interact with the redirected page.

After this commit:

The conversation is automatically folded when the chatbot redirects a
visitor to a webpage on mobile.

Task-4630719

closes odoo/odoo#215138

X-original-commit: 8a824a73b62c933a89de2e4660f852c2c66dfbe2
Signed-off-by: Matthieu Stockbauer (tsm) <tsm@odoo.com>